+++ unstructured/utils.py
@@ -5,8 +5,10 @@ import json
 import os
 import platform
 import subprocess
+import sys
 from datetime import datetime
 from functools import wraps
+from glob import glob
 from itertools import combinations
 from typing import (
     TYPE_CHECKING,
@@ -719,3 +721,64 @@ def catch_overlapping_and_nested_bboxes(
                 document_with_overlapping_flag = True
 
     return document_with_overlapping_flag, overlapping_cases
+
+
+def get_homedir() -> str:
+    if sys.platform == "win32" and "APPDATA" in os.environ:
+        homedir = os.environ["APPDATA"]
+    else:
+        homedir = os.path.expanduser("~/")
+        if homedir == "~/":
+            raise ValueError("Could not find a default directory for TreeSitter")
+    return homedir
+
+
+def update_tree_sitter(languages: List[str], cache_dir: Optional[str] = None):
+    if dependency_exists("tree_sitter"):
+        from tree_sitter import Language
+    else:
+        raise RuntimeError("Cannot update TreeSitter if not installed")
+
+    if not cache_dir:
+        cache_dir = get_homedir()
+    else:
+        raise NotImplementedError("Custom treesitter cache directory is not available yet")
+
+    download_dir = os.path.join(cache_dir, ".unstructured_treesitter")
+
+    installed_languages = [
+        l.split("/")[-1].removeprefix("tree-sitter-") for l in glob(f"{download_dir}/tree-sitter-*")
+    ]
+
+    available_parsers = []
+    unavailable_parsers = []
+    for language in languages:
+        if language in installed_languages:
+            continue
+        if (
+            requests.get(f"https://github.com/tree-sitter/tree-sitter-{language}").status_code
+            == 200
+        ):
+            available_parsers.append(language)
+        else:
+            unavailable_parsers.append(language)
+
+    if len(unavailable_parsers) > 0:
+        raise ValueError(f"Treesitter parsers are unavailable for: {','.join(unavailable_parsers)}")
+
+    for language in available_parsers:
+        subprocess.run(
+            f"git clone https://github.com/tree-sitter/tree-sitter-{language} {download_dir}/tree-sitter-{language}",
+            shell=True,
+        )
+
+    all_languages = glob(f"{download_dir}/tree-sitter-*/**/src/parser.c", recursive=True)
+    # Some languages might have different folder structure
+    all_languages = [l.removesuffix("/src/parser.c") for l in all_languages]
+
+    # Update main shared library
+    Language.build_library(f"{download_dir}/treesitter_build/languages_tmp.so", all_languages)
+    subprocess.run(
+        f"mv {download_dir}/treesitter_build/languages_tmp.so {download_dir}/treesitter_build/languages.so",
+        shell=True,
+    )