+++ CHANGELOG.md
@@ -71,6 +71,7 @@
 ### Features
 
 * **Add Databricks Volumes destination connector** Databricks Volumes connector added to ingest CLI.  Users may now use `unstructured-ingest` to write partitioned data to a Databricks Volumes storage service.
+* **Code Partition.** Partition code file using TreeSitter instead of text partition for the following languages : Python Javascript Typescript C C++ Go Ruby Swift C# PHP 
 
 ### Fixes